refactor: assets 폴더 정리
